feat: add missing old validation tests to textarea
