+++ packages/web/src/components/gcds-textarea/test/gcds-textarea.e2e.html
@@ -33,6 +33,7 @@
         label="Label"
         name="textarea-name"
         hint="Hint message."
+        required
       >
       </gcds-textarea>
     </form>
